Separated the tests into internal and public
After separating them, I moved some references and started using sinon
for mocking instead of having to use some internal function. Tests pass
and they should be good for min-max testing, etc.
