+++ sdk/identity/identity/package.json
@@ -136,6 +136,8 @@
     "rollup-plugin-terser": "^5.1.1",
     "rollup-plugin-visualizer": "^4.0.4",
     "typescript": "~3.9.3",
-    "util": "^0.12.1"
+    "util": "^0.12.1",
+    "sinon": "^9.0.2",
+    "@types/sinon": "^9.0.4"
   }
 }
