+++ sdk/identity/identity/test/public/azureCliCredential.spec.ts
