+++ sdk/identity/identity/test/public/clientCertificateCredential.spec.ts
