+++ sdk/identity/identity/test/public/clientSecretCredential.spec.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { ClientSecretCredential } from "../src";
-import { MockAuthHttpClient, assertClientCredentials } from "./authTestUtils";
+import { ClientSecretCredential } from "../../src";
+import { MockAuthHttpClient, assertClientCredentials } from "../authTestUtils";
 
 describe("ClientSecretCredential", function() {
   it("sends an authorization request with the given credentials", async () => {
