+++ sdk/identity/identity/test/public/errors.spec.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import assert from "assert";
-import { AggregateAuthenticationError } from "../src";
+import { AggregateAuthenticationError } from "../../src";
 
 describe("AggregateAuthenticationError", function() {
   it("produces a message containing details of the errors it contains", async () => {
